+++ CHANGELOG.md
@@ -5,6 +5,7 @@ ChangeLog
 0.9.4 (2025-11-25)
 ------------------
 
+- IMPROVEMENT: allow docker(1) compatible command to be explicity set, too
 - IMPROVEMENT: blindly add support for Podman [Desktop] and Rancher Desktop
 
 0.9.3 (2025-11-25)